Most IPC implemented

+++ src/c/autonomous_trust/utilities/util.c
@@ -22,9 +22,9 @@
 
 #include "util.h"
 
-inline int min(int a, int b) { return ((a) < (b) ? a : b); }
+inline long min(long a, long b) { return ((a) < (b) ? a : b); }
 
-inline int max(int a, int b) { return ((a) > (b) ? a : b); }
+inline long max(long a, long b) { return ((a) > (b) ? a : b); }
 
 char *strremove(char *str, const char *sub)
 {
